Singleton - konstruktor hiba javívta A vizualizációs csapat kérésére gyűjtő osztályok létrehozva
