+++ DataBus/src/Environment/misc/Misc.java
@@ -1,0 +1,12 @@
+package Environment.misc;
+
+import Environment.WorldObject;
+
+/**
+ * Created by Németh Kriszitna on 2016.11.03..
+ */
+public abstract class Misc extends WorldObject {
+    public Misc(int Id, int[] startPosition, int width, int height, double[] Transform, int Zlevel, int Opacity, boolean CanStuckOnIt) {
+        super(Id, startPosition, width, height, Transform, Zlevel, Opacity, CanStuckOnIt);
+    }
+}
