+++ DataBus/src/Environment/misc/Parking.java
@@ -7,9 +7,10 @@ import java.util.Arrays;
 /**
  * Created by Akos on 2016. 09. 27..
  */
-public class Parking extends WorldObject{
-    public Parking(int Id, int[] startPosition, int Width, int Height, double[] Transform, int Zlevel, int Opacity, ParkingElement parkingElement, boolean CanStuckOnIt) {
-        super(Id, startPosition, Width, Height, Transform, Zlevel, Opacity, CanStuckOnIt);
+public class Parking extends Misc{
+
+    public Parking(int Id, int[] startPosition, int width, int height, double[] Transform, int Zlevel, int Opacity, ParkingElement parkingElement, boolean CanStuckOnIt) {
+        super(Id, startPosition, width, height, Transform, Zlevel, Opacity, CanStuckOnIt);
         this.parkingElement = parkingElement;
     }
