+++ DataBus/src/Environment/misc/Tree.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 /**
  * Created by nemeth on 2016. 09. 30..
  */
-public class Tree extends WorldObject{
+public class Tree extends Misc{
     public Tree(int Id, int[] startPosition, double[] Transform, int Zlevel, int Opacity) {
         super(Id, startPosition, 145, 160, Transform, Zlevel, Opacity, true);
     }
