+++ DataBus/src/Environment/road_signs/Priority.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 /**
  * Created by nemeth on 2016. 09. 30..
  */
-public class Priority extends WorldObject {
+public class Priority extends RoadSign {
     public Priority(int Id, int[] startPosition, double[] Transform, int Zlevel, int Opacity, PriorityType priorityType) {
         super(Id, startPosition, 80, 80, Transform, Zlevel, Opacity, false);
         this.priorityType = priorityType;
