+++ DataBus/src/Environment/road_tiles/LaneAdvanced.java
@@ -7,30 +7,16 @@ import java.util.Arrays;
 /**
  * Created by nemeth on 2016. 09. 30..
  */
-public class LaneAdvanced extends WorldObject {
-    public enum RoadPaintings1 {rp_none,rp_sep_1_s}
-    public enum RoadPaintings2 {rp_none,rp_sep_2_s}
-    public enum RoadPaintings3 {rp_none,rp_sep_3_s}
-    public enum LaneAdvancedType {CrossRoads,Rotary,TJunctionLeft,TJunctionRight}
+public class LaneAdvanced extends RoadTile {
 
-    RoadPaintings1 roadPaintings1;
-    RoadPaintings2 roadPaintings2;
-    RoadPaintings3 roadPaintings3;
-    LaneAdvancedType laneAdvancedType;
-    int[] roadColor1;
-    int[] roadColor2;
-    int[] roadColor3;
 
-    public LaneAdvanced(int Id, int[] startPosition, int Width, int Height, double[] Transform, int Zlevel, int Opacity, int[] roadColor1, int[] roadColor2, int[] roadColor3, RoadPaintings1 roadPaintings1, RoadPaintings2 roadPaintings2, RoadPaintings3 roadPaintings3, LaneAdvancedType laneAdvancedType) {
-        super(Id, startPosition, Width, Height, Transform, Zlevel, Opacity, false);
-        this.roadPaintings1 = roadPaintings1;
-        this.roadPaintings2 = roadPaintings2;
-        this.roadPaintings3 = roadPaintings3;
+    public LaneAdvanced(int Id, int[] startPosition, int width, int height, double[] Transform, int Zlevel, int Opacity, int[] roadColor1, int[] roadColor2, int[] roadColor3, RoadPaintings1 roadPaintings1, RoadPaintings2 roadPaintings2, RoadPaintings3 roadPaintings3, LaneAdvancedType laneAdvancedType) {
+        super(Id, startPosition, width, height, Transform, Zlevel, Opacity, false, roadColor1, roadColor2, roadColor3, roadPaintings1, roadPaintings2, roadPaintings3);
         this.laneAdvancedType = laneAdvancedType;
-        this.roadColor1 = roadColor1;
-        this.roadColor2 = roadColor2;
-        this.roadColor3 = roadColor3;
     }
+
+    public enum LaneAdvancedType {CrossRoads,Rotary,TJunctionLeft,TJunctionRight}
+    LaneAdvancedType laneAdvancedType;
 
     @Override
     public String toString() {
