+++ DataBus/src/Environment/road_tiles/LaneSimple.java
@@ -7,47 +7,17 @@ import java.util.Arrays;
 /**
  * Created by nemeth on 2016. 09. 30..
  */
-public class LaneSimple extends WorldObject{
-    public enum RoadPaintings1 {rp_none,rp_sep_1_s,rp_sep_1_sb}
-    public enum RoadPaintings2 {rp_none,rp_sep_2_bnd,rp_sep_2_d,rp_sep_2_dbnd,rp_sep_2_ds,rp_sep_2_s,rp_sep_2_sd,rp_sep_2_ss}
-    public enum RoadPaintings3 {rp_none,rp_sep_3_s,rp_sep_3_sb}
+public class LaneSimple extends RoadTile{
+
+    public LaneSimple(int Id, int[] startPosition, int width, int height, double[] Transform, int Zlevel, int Opacity, int[] roadColor1, int[] roadColor2, int[] roadColor3, RoadPaintings1 roadPaintings1, RoadPaintings2 roadPaintings2, RoadPaintings3 roadPaintings3, LaneSimpleType laneSimpleType) {
+        super(Id, startPosition, width, height, Transform, Zlevel, Opacity, false, roadColor1, roadColor2, roadColor3, roadPaintings1, roadPaintings2, roadPaintings3);
+        this.laneSimpleType = laneSimpleType;
+    }
+
     public enum LaneSimpleType {Left45,Left65,Left90,Right45,Right65,Right90,Straight}
-
-    // rp_sep_1 bal szél
-    // rp_sep_2 közép
-    // rp_sep_3 jobb szél
-    // none - nincs
-    // s - sima vonal (vékony)
-    // sb - sima vonal (vastag - bold)
-    // d - szaggatott vonal
-    // ss - dupla sima vonal (dupla záróvonal)
-    // ds - szaggatott záró
-    // sd - záró szaggatott
-    // bnd
-    // dbnd - double pontozott (talán a fényvisszaverő)
-
-    int[] roadColor1;
-    int[] roadColor2;
-    int[] roadColor3;
-
-    RoadPaintings1 roadPaintings1;
-    RoadPaintings2 roadPaintings2;
-    RoadPaintings3 roadPaintings3;
     LaneSimpleType laneSimpleType;
 
-    public LaneSimple(int Id, int[] startPosition, int Width, int Height,
-                      double[] Transform, int Zlevel, int Opacity, int[] roadColor1, int[] roadColor2, int[] roadColor3, RoadPaintings1 roadPaintings1, RoadPaintings2 roadPaintings2,
-                      RoadPaintings3 roadPaintings3, LaneSimpleType laneSimpleType) {
-        super(Id, startPosition, Width, Height, Transform, Zlevel, Opacity, false);
 
-        this.roadPaintings1 = roadPaintings1;
-        this.roadPaintings2 = roadPaintings2;
-        this.roadPaintings3 = roadPaintings3;
-        this.laneSimpleType = laneSimpleType;
-        this.roadColor1 = roadColor1;
-        this.roadColor2 = roadColor2;
-        this.roadColor3 = roadColor3;
-    }
 
     @Override
     public String toString() {
